display refactor - kamangir/bolt#703
