+++ blue_sbc/imager/__init__.py
@@ -1,10 +1,15 @@
 from abcli.modules.cookie import cookie
+import abcli.logging
+import logging
+
+logger = logging.getLogger(__name__)
 
 NAME = "blue_sbc.imager"
 
 from .classes import *
 
-if cookie.get("session.imager") == "lepton":
+imager_name = cookie.get("session.imager", "camera")
+if imager_name == "lepton":
     from .lepton import instance as imager
 else:
     from .camera import instance as imager
