+++ blue_sbc/screen/__init__.py
@@ -1,0 +1,13 @@
+from abcli.modules.cookie import cookie
+import abcli.logging
+import logging
+
+logger = logging.getLogger(__name__)
+
+NAME = "blue_sbc.screen"
+
+# screen_name = cookie.get("session.screen", "display")
+# if screen_name == "display":
+from .display import instance as screen
+
+logger.info(f"{NAME}: imager: {screen.__class__.__name__}")
