+++ blue_sbc/screen/classes.py
@@ -1,0 +1,2 @@
+class Screen(object):
+    ...
