+++ blue_sbc/screen/display/__init__.py
@@ -1,4 +1,4 @@
-NAME = "blue_sbc.display"
+NAME = "blue_sbc.screen.display"
 
 from .classes import Display
